+++ src/index.js
@@ -4,7 +4,6 @@
 import './_style/index.scss'
 
 export Alert from './Alert'
-export Dialog from './Dialog'
 export Button from './Button'
 export Cell from './Cell/Cell'
 export CellBody from './Cell/CellBody'
@@ -14,9 +13,11 @@ export CellTip from './Cell/CellTip'
 export Cells from './Cell/Cells'
 export CellInput from './CellInput'
 export DatePicker from './DatePicker'
+export Modal from './Modal'
+export Dialog from './Dialog'
 export Form from './Form/Form'
-export Input from './Form/Input'
 export CheckBox from './Form/CheckBox'
+export Input from './Form/Input'
 export Select from './Form/Select'
 export Icon from './Icon'
 export Panel from './Panel'
